# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EvidenceType": {
        "description": "Evidence source types permitted by the frozen v1.0 scenario.",
        "enum": [
          "DOCUMENT",
          "DATABASE",
          "CALCULATION"
        ],
        "title": "EvidenceType",
        "type": "string"
      },
...
      "JsonValue": {},
...
      "NaturalLanguageTaskSubmission": {
        "additionalProperties": false,
        "description": "Public request: natural-language task plus tightening-only execution options.",
        "properties": {
          "max_steps": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Metadata",
            "type": "object"
          },
          "output_format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskOutputFormat"
              },
              {
                "type": "null"
              }
            ]
          },
          "read_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Only"
          },
          "require_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Require Approval"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "task": {
            "minLength": 1,
            "title": "Task",
            "type": "string"
          },
          "task_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "task"
        ],
        "title": "NaturalLanguageTaskSubmission",
        "type": "object"
      },
...
      "RuntimeStatus": {
        "description": "Execution-host status kept separate from the frozen business TaskStatus.",
        "enum": [
          "READY",
          "LEASED",
          "WAITING_RETRY",
          "SUSPENDED",
          "FINISHED"
        ],
        "title": "RuntimeStatus",
        "type": "string"
      },
      "TaskErrorResponse": {
        "additionalProperties": false,
        "description": "Uniform transport error returned before a workflow result exists.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message",
          "trace_id"
        ],
        "title": "TaskErrorResponse",
        "type": "object"
      },
      "TaskEvidenceListResponse": {
        "additionalProperties": false,
        "description": "Deterministically ordered task Evidence collection.",
        "properties": {
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/TaskEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "evidence"
        ],
        "title": "TaskEvidenceListResponse",
        "type": "object"
      },
      "TaskEvidenceResponse": {
        "additionalProperties": false,
        "description": "Minimized Evidence metadata with retained lineage.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "content_summary": {
            "title": "Content Summary",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Source"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "formula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formula"
          },
          "input_evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Input Evidence Ids",
            "type": "array"
          },
          "lineage": {
            "items": {
              "type": "string"
            },
            "title": "Lineage",
            "type": "array"
          },
          "produced_by": {
            "title": "Produced By",
            "type": "string"
          },
          "query_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Id"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EvidenceType"
          }
        },
        "required": [
          "evidence_id",
          "type",
          "source",
          "produced_by",
          "step_id",
          "lineage",
          "confidence",
          "created_at",
          "query_id",
          "document_source",
          "formula",
          "input_evidence_ids",
          "content_summary"
        ],
        "title": "TaskEvidenceResponse",
        "type": "object"
      },
      "TaskListResponse": {
        "additionalProperties": false,
        "description": "Bounded current-user task history ordered newest-first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskOutputFormat": {
        "description": "Public output choices mapped onto the frozen Artifact contract.",
        "enum": [
          "pdf",
          "json"
        ],
        "title": "TaskOutputFormat",
        "type": "string"
      },
      "TaskResponse": {
        "additionalProperties": false,
        "description": "Stable public task-management summary.",
        "properties": {
          "artifact_count": {
            "title": "Artifact Count",
            "type": "integer"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "pending_approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Approval Id"
          },
          "runtime_status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "step_count": {
            "title": "Step Count",
            "type": "integer"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_summary": {
            "title": "Task Summary",
            "type": "string"
          },
          "task_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskType"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "trace_id",
          "status",
          "runtime_status",
          "task_type",
          "created_at",
          "started_at",
          "completed_at",
          "cancelled_at",
          "current_step",
          "task_summary",
          "pending_approval_id",
          "step_count",
          "evidence_count",
          "artifact_count",
          "error_summary"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "TaskStatus": {
        "description": "Authoritative lifecycle states from the frozen task state machine.",
        "enum": [
          "CREATED",
          "UNDERSTANDING",
          "PLANNING",
          "EXECUTING",
          "WAITING_APPROVAL",
          "RETRYING",
          "REPLANNING",
          "VERIFYING",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "TaskStepsResponse": {
        "additionalProperties": false,
        "description": "Deterministically ordered step collection.",
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TaskStepResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "steps"
        ],
        "title": "TaskStepsResponse",
        "type": "object"
      },
      "TaskSubmissionResponse": {
        "additionalProperties": false,
        "description": "Future 202 Accepted response returned after Task and dispatch commit atomically.",
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "artifacts_url": {
            "minLength": 1,
            "title": "Artifacts Url",
            "type": "string"
          },
          "runtime_status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "status_url": {
            "minLength": 1,
            "title": "Status Url",
            "type": "string"
          },
          "task_id": {
            "minLength": 1,
            "title": "Task Id",
            "type": "string"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "trace_id": {
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "trace_id",
          "task_status",
          "runtime_status",
          "accepted_at",
          "status_url",
          "artifacts_url"
        ],
        "title": "TaskSubmissionResponse",
        "type": "object"
      },
      "TaskType": {
        "description": "Versioned business task classifications admitted by governed contracts.",
        "enum": [
          "supplier_quality_analysis.v1",
          "accounts_payable_analysis.v1"
        ],
        "title": "TaskType",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Agentic Enterprise Knowledge Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tasks": {
      "get": {
        "description": "Return one bounded tenant- and owner-scoped task history page.",
        "operationId": "list_tasks",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "description": "Durably accept one Task without executing LangGraph in the API process.",
        "operationId": "create_task",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NaturalLanguageTaskSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Submit Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "description": "Return an authorized task summary through the application service.",
        "operationId": "get_task",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/v1/tasks/{task_id}/cancel": {
      "post": {
        "description": "Durably accept cancellation; underlying I/O may stop at a later safe boundary.",
        "operationId": "cancel_task",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Cancel Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/evidence": {
      "get": {
        "description": "Return only persisted, minimized Evidence metadata and lineage.",
        "operationId": "list_task_evidence",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEvidenceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Task Evidence",
        "tags": [
          "tasks"
        ]
      }
    },
    "/v1/tasks/{task_id}/steps": {
      "get": {
        "description": "Return planned steps combined with already persisted results.",
        "operationId": "list_task_steps",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStepsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Task Steps",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}